Fixed the end game clear issue, rebalanced level one starting diffiuclty, started factory coreography.
In the factory scene i tried some coreography stuff to mimik the script. Biggest change was finally solving the clear on fial issue. Now whn a level ends instead of cutting cpawning it simply scales it down over time. When you fail instead of clearing birds it disables them for a short period but keeps them present so as to not cause a "clear" untill a set delay.